+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/netatmo/NetatmoService.kt
@@ -3,8 +3,6 @@ package org.agrfesta.sh.api.providers.netatmo
 import arrow.core.Either
 import kotlinx.coroutines.runBlocking
 import org.agrfesta.sh.api.core.application.ports.outbounds.devices.DevicesProvider
-import org.agrfesta.sh.api.core.domain.devices.DeviceFeature.ACTUATOR
-import org.agrfesta.sh.api.core.domain.devices.DeviceFeature.SENSOR
 import org.agrfesta.sh.api.core.domain.devices.DeviceModel
 import org.agrfesta.sh.api.core.domain.devices.Provider
 import org.agrfesta.sh.api.core.domain.devices.Provider.NETATMO
@@ -40,7 +38,6 @@ class NetatmoService(
                             deviceProviderId = module.id,
                             provider = provider,
                             name = module.name,
-                            features = setOf(SENSOR, ACTUATOR),
                             model = DeviceModel(SMARTHER_MODEL)
                         )
                     }